+++ src/scenes/board/index.tsx
@@ -202,20 +202,6 @@ const StyledControls = styled.div`
   right:0;
 `;
 
-const StyledDebugThing = styled.div`
-  position: absolute;
-  top:2rem;
-  left:1rem;
-
-  font-size: 3rem;
-  text-align:left;
-  color:white;
-  pointer-events: none;
-`;
-
-const evCache: any[] = [];
-let prevDiff: number = -1;
-
 export function Board() {
   const dispatch = useAppDispatch();
   const grid = useAppSelector(selectGridBox);
@@ -223,7 +209,6 @@ export function Board() {
   const gridInfo = useAppSelector(selectGridInfo);
   const zoom = useAppSelector(getZoom);
   const position = useAppSelector(getPosition);
-  const [ debugMessage, setDebugMessage ] = useState('');
 
   const onClickCell = useCallback((cellIdx) => {
     dispatch(rotateCell(cellIdx));
@@ -238,67 +223,6 @@ export function Board() {
     return `${Math.round(100 / gridInfo.numValues)}%`;
   }, [ gridInfo.numValues ])
 
-  const onPointerCancel = useCallback((ev) => {
-    // Remove this pointer from the cache
-    const foundIdx = evCache.findIndex(evC => ev.pointerId === evC.pointerId);
-    if(foundIdx > -1) {
-      console.log('cancelling ', evCache[foundIdx].pointerId);
-      evCache.splice(foundIdx, 1);
-    }
-  
-    // If the number of pointers down is less than two then reset diff tracker
-    if (evCache.length < 2) {
-      prevDiff = -1;
-      setDebugMessage('CANCELLED');
-    }
-  }, [setDebugMessage]);
-
-  const onPointerDown = useCallback((ev) => {
-    // The pointerdown event signals the start of a touch interaction.
-    // This event is cached to support 2-finger gestures
-    evCache.push(ev);
-    // log("pointerDown", ev);
-  }, []);
-
-  const onPointerMove = useCallback((ev) => {
-    // example from https://developer.mozilla.org/en-US/docs/Web/API/Pointer_events/Pinch_zoom_gestures
-    // This function implements a 2-pointer horizontal pinch/zoom gesture.
-    //
-    // If the distance between the two pointers has increased (zoom in),
-    // the target element's background is changed to "pink" and if the
-    // distance is decreasing (zoom out), the color is changed to "lightblue".
-    //
-    // This function sets the target element's border to "dashed" to visually
-    // indicate the pointer's target received a move event.
-    console.log('onPointerMove', ev.pointerId);
-    console.log(evCache);
-
-    // Find this event in the cache and update its record with this event
-    const index = evCache.findIndex((cachedEv) => cachedEv.pointerId === ev.pointerId);
-    evCache[index] = ev;
-
-    // If two pointers are down, check for pinch gestures
-    if (evCache.length === 2) {
-      // Calculate the distance between the two pointers
-      const curDiff = Math.abs(evCache[0].clientX - evCache[1].clientX);
-
-      if (prevDiff > 0) {
-        if (curDiff > prevDiff) {
-          // The distance between the two pointers has increased
-          console.log('Pinch moving OUT -> Zoom in', ev);
-          setDebugMessage(`OUT (${curDiff})`);
-        }
-        if (curDiff < prevDiff) {
-          // The distance between the two pointers has decreased
-          console.log('Pinch moving IN -> Zoom out',ev);
-          setDebugMessage(`IN (${curDiff})`);
-        }
-      }
-
-      // Cache the distance for the next move event
-      prevDiff = curDiff;
-    }
-  }, [ setDebugMessage ]);
 
   const renderCellGroup = (cellGroup: CellObj[], cgKey: string, gridSize: number, cellRatio: string, boardCell: RawCell) => {
     // [0] check here cause this is all janky and the individual cells are undefined on load
@@ -335,18 +259,8 @@ export function Board() {
   };
 
   return (
-    <StyledBoardContainer 
-      onPointerDown={onPointerDown}
-      onPointerMove={onPointerMove}
-      onPointerUp={onPointerCancel}
-      onPointerCancel={onPointerCancel}
-      onPointerOut={onPointerCancel}
-      onPointerLeave={onPointerCancel}
-      >
+    <StyledBoardContainer>
       <PositionControls />
-      <StyledDebugThing>
-        {`DEBUG: ${debugMessage}`}
-      </StyledDebugThing>
       <StyledBoard style={tStyles}>
         <StyledControls>
           <BoardControls />
